Update project 2
Hi! I changed the structure of the code a bit, now all the physical quantities as calculated in functions and no longer in the main run. In this way I can also implement the finite size scaling in the original program (I think), I am now going to work on that.